pyproject.toml
[project]
name = "crm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "celery-types>=0.26.0",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.2",
    "django-modeltranslation>=0.20.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.4.0",
    "reportlab>=4.4.10",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.11.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings_test"
python_files  = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=apps --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["apps"]
omit   = ["*/migrations/*", "*/tests/*", "*/admin.py"]
